Refactor state resources persistence code to remove abstractions
