+++ state/state.go
@@ -1344,11 +1344,8 @@ func (st *State) AddApplication(args AddApplicationArgs) (_ *Application, err er
 
 		if len(args.Resources) > 0 {
 			// Collect pending resource resolution operations.
-			resources, err := st.Resources()
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			resOps, err := resources.NewResolvePendingResourcesOps(args.Name, args.Resources)
+			resources := st.Resources().(*resourcePersistence)
+			resOps, err := resources.resolveApplicationPendingResourcesOps(args.Name, args.Resources)
 			if err != nil {
 				return nil, errors.Trace(err)
 			}
