+++ testing/factory/factory.go
@@ -472,7 +472,7 @@ func (factory *Factory) MakeApplication(c *gc.C, params *ApplicationParams) *sta
 	return app
 }
 
-// MakeApplication creates an application with the specified parameters, substituting
+// MakeApplicationReturningPassword creates an application with the specified parameters, substituting
 // sane defaults for missing values.
 // If params is not specified, defaults are used.
 // It returns the application and its password.
@@ -492,8 +492,7 @@ func (factory *Factory) MakeApplicationReturningPassword(c *gc.C, params *Applic
 		c.Assert(err, jc.ErrorIsNil)
 	}
 
-	rSt, err := factory.st.Resources()
-	c.Assert(err, jc.ErrorIsNil)
+	rSt := factory.st.Resources()
 
 	resourceMap := make(map[string]string)
 	for name, res := range params.Charm.Meta().Resources {
@@ -568,7 +567,7 @@ func (factory *Factory) MakeUnit(c *gc.C, params *UnitParams) *state.Unit {
 	return unit
 }
 
-// MakeUnit creates an application unit with specified params, filling in sane
+// MakeUnitReturningPassword creates an application unit with specified params, filling in sane
 // defaults for missing values. If params is not specified, defaults are used.
 // The unit and its password are returned.
 //
